feat: change metadata entry to use Double instead of Long

+++ webapps-schema/src/main/java/io/camunda/webapps/schema/entities/clustervariable/ClusterVariableEntity.java
@@ -160,5 +160,5 @@ public class ClusterVariableEntity implements ExporterEntity<ClusterVariableEnti
         + '}';
   }
 
-  public record MetadataEntry(String key, String value, Long valueNumber) {}
+  public record MetadataEntry(String key, String value, Double valueNumber) {}
 }
